+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Fixed an issue where `SExp.to(<boolean>)` did not work the same as python's clvm.
 - Fixed an issue where Bytes-SExp comparison like `b('aaa').equal_to(SExp.null())` did not work.
+- Fixed an issue where `OperatorDict` did not throw Error when either `quote` or `apply` is not specified.
 ### Added
 - Added `list()` function working similar to Python's `list()`.